% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_engine.R
\name{get_similar_players}
\alias{get_similar_players}
\title{Find similar players using cosine similarity}
\usage{
get_similar_players(model, player_name, top_n = 5)
}
\arguments{
\item{model}{A trained scout brain model}

\item{player_name}{Character. Name of the reference player}

\item{top_n}{Integer. Number of similar players to return}
}
\value{
A tibble of similar players
}
\description{
Find similar players using cosine similarity
}
