% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_real_scout.R
\name{save_raw_data}
\alias{save_raw_data}
\title{Save raw scouting data to CSV}
\usage{
save_raw_data(df)
}
\arguments{
\item{df}{A data frame containing raw scouting results}
}
\value{
File path of saved CSV
}
\description{
Save raw scouting data to CSV
}
