% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_real_scout.R
\name{scrape_club}
\alias{scrape_club}
\title{Scrape all players from a club}
\usage{
scrape_club(club_url, league_country)
}
\arguments{
\item{club_url}{URL of the club page}

\item{league_country}{Character. League or country name}
}
\value{
A tibble of players
}
\description{
Scrape all players from a club
}
