% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_real_scout.R
\name{scrape_league}
\alias{scrape_league}
\title{Scrape one league from Transfermarkt}
\usage{
scrape_league(league_url, league_country = "Unknown League")
}
\arguments{
\item{league_url}{URL liga Transfermarkt}

\item{league_country}{Nama liga / negara}
}
\value{
tibble
}
\description{
Scrape one league from Transfermarkt
}
