% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_real_scout.R
\name{scrape_player}
\alias{scrape_player}
\title{Scrape a single player row}
\usage{
scrape_player(node)
}
\arguments{
\item{node}{HTML node corresponding to a player row}
}
\value{
A tibble with player information
}
\description{
Scrape a single player row
}
