% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_engine.R
\name{train_scout_brain}
\alias{train_scout_brain}
\title{Train similarity-based scout model}
\usage{
train_scout_brain(df)
}
\arguments{
\item{df}{A cleaned player data frame}
}
\value{
A scout brain model object
}
\description{
Train similarity-based scout model
}
